+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # gasket
 
+- Set version of mermaid and docsify for gasket.dev. ([#664])
+
 - #[#128], [#131] Update to latest `docsify`
 
 ## 5.0.0
@@ -12,4 +14,4 @@
 [#101]: https://github.com/godaddy/gasket/pull/101
 [#128]: https://github.com/godaddy/gasket/issues/128
 [#131]: https://github.com/godaddy/gasket/pull/131
-
+[#664]: https://github.com/godaddy/gasket/pull/664